add group-exchange kex (key exchange). this is a more complicated way of doing the initial DH key exchange, where the client asks the server for a modulus of a certain size and the server picks an appropriate one at random from a set. theoretically this is more secure than normal "group 1" key exchange, but i'm dubious. anyway, it's included for completeness.
i've included the modulus file with a bunch of moduli from openssh. this
should work fine for any standard JVM, but on an embedded device, it will
probably fail, because it relies on loading files from the jar via the
class loader. in this case it will just silently fail to load the file (i'm
using reflection tricks) and not advertise gex.